requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
-e git://github.com/django/django.git@afeafd6036616bac8263d762c1610f22241c0187#egg=Django
django-heroku==0.3.1
gunicorn==20.1.0
mysqlclient==2.0.3
psycopg2==2.9.2
sqlparse==0.4.2
whitenoise==5.3.0
